refactor(ui): rename type files with descriptive names instead of index.ts
