+++ ui/actions/overview/findings/types/index.ts
@@ -1,25 +1,1 @@
-// Findings Severity Overview Types
-// Corresponds to the /overviews/findings_severity endpoint
-
-interface OverviewResponseMeta {
-  version: string;
-}
-
-export interface FindingsSeverityAttributes {
-  critical: number;
-  high: number;
-  medium: number;
-  low: number;
-  informational: number;
-}
-
-export interface FindingsSeverityOverview {
-  type: "findings-severity-overview";
-  id: string;
-  attributes: FindingsSeverityAttributes;
-}
-
-export interface FindingsSeverityOverviewResponse {
-  data: FindingsSeverityOverview;
-  meta: OverviewResponseMeta;
-}
+export * from "./findings.types";
